Bring outline view branch up to date with trunk (r3203:r3224).

+++ Source/SPArrayAdditions.h
@@ -24,7 +24,7 @@
 
 static inline id NSArrayObjectAtIndex(NSArray *self, NSUInteger i) 
 {
-	return (id)CFArrayGetValueAtIndex((CFArrayRef)self, i);
+	return (id)CFArrayGetValueAtIndex((CFArrayRef)self, (long)i);
 }
 
 static inline void NSMutableArrayAddObject(NSArray *self, id anObject)
